refactor(core): split coord geometry and pipeline coord transform tests
Extract tessellation budget/segment math and path projection from
coord-geometry.ts; keep projectGeometryBatch as the stable dispatch
facade with re-exported budget helpers for existing import paths.

Replace the pipeline-coord-transform characterization monolith with
focused suites under tests/pipeline-coord-transform/.
